extruct sfn code and state machine code, separate file
